requirements.txt
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-allauth==0.63.3
django-cleanup==8.1.0
django-htmx==1.17.3
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
sqlparse==0.5.0
yarl==1.9.4
